requirements.txt
langgraph>=0.2.0
gspread>=6.0.0
google-auth>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
